requirements.txt
langchain>=0.2.1
langchain-openai>=0.2.1
langchain-google-genai>=0.0.3
openai>=1.1.0
google-generativeai>=0.8.0
anthropic>=0.7.0
PyGithub>=2.1.0